tick logic inside machine
